Fix compilation for linux

+++ src/NotepadNext/LanguageStylesModel.cpp
@@ -117,14 +117,14 @@ QVariant LanguageStylesModel::data(const QModelIndex &index, int role) const
             case 4: // Font
                 return editor->styleFont(index.row());
             case 5: // Size
-                return editor->styleSize(index.row());
+                return (int)editor->styleSize(index.row());
             case 6: // Fractional
-                return editor->styleSizeFractional(index.row());
+                return (int)editor->styleSizeFractional(index.row());
             case 7: // Bold
                 if (role == Qt::EditRole) return editor->styleBold(index.row());
                 else return editor->styleBold(index.row()) ? "True" : "False";
             case 8: // Weight
-                return editor->styleWeight(index.row());
+                return (int)editor->styleWeight(index.row());
             case 9: // Italic
                 if (role == Qt::EditRole) return editor->styleItalic(index.row());
                 else return editor->styleItalic(index.row()) ? "True" : "False";
@@ -139,9 +139,9 @@ QVariant LanguageStylesModel::data(const QModelIndex &index, int role) const
                 if (role == Qt::EditRole) return editor->styleEOLFilled(index.row());
                 else return editor->styleEOLFilled(index.row()) ? "True" : "False";
             case 14: // Character Set
-                return editor->styleCharacterSet(index.row());
+                return (int)editor->styleCharacterSet(index.row());
             case 15: // Case
-                if (role == Qt::EditRole) return editor->styleCase(index.row());
+                if (role == Qt::EditRole) return (int)editor->styleCase(index.row());
                 else return val_to_case_str(editor->styleCase(index.row()));
             case 16: // Visible
                 if (role == Qt::EditRole) return editor->styleVisible(index.row());
